refactor(js-paths): align import statements with new modular file structure
