+++ projekt/public/html/dashboard.html
@@ -42,7 +42,7 @@
 			
 		</p>
 		
-		<script	type="module"	src="/js/dashboard.js"></script>
+		<script	type="module"	src="/js/dashboard/index.js"></script>
 		<script	src="/js/logout.js"></script>
 	</body>
 </html>
