+++ projekt/public/js/dashboard/index.js
@@ -1,11 +1,11 @@
 // Funktionale-Imports
-import {getTodos, addTodo, toggleTodoStatus, deleteTodo} from "./api/todo.js"
-import {getUserInfo} from "./api/user.js"
+import {getTodos, addTodo, toggleTodoStatus, deleteTodo} from "./../shared/api/todo.js"
+import {getUserInfo} from "./../shared/api/user.js"
 
 // Import von Selektoren
-import {dashboardSelectors} from "./dom/dashboard/selectors.js"
+import {dashboardSelectors} from "./dom/selectors.js"
 // Import Todolistengenerierung
-import {createTodoListItem, createEmptyTodoMessage} from "./dom/dashboard/create.js"
+import {createTodoListItem, createEmptyTodoMessage} from "./dom/create.js"
 
 document.addEventListener("DOMContentLoaded", () => {
 	/*
