+++ projekt/public/js/shared/api/todo.js
@@ -1,7 +1,7 @@
 // Importiert den zentralen fetch()-Wrapper für standardisierte HTTP-Anfragen
 import {apiRequest} from "./fetchWrapper.js"
 // Importiert Hilfsfunktion zur Token-Ermittlung aus dem localStorage
-import {getToken} from "../utils/token.js"
+import {getToken} from "./../utils/token.js"
 
 // Ruft alle Todos des aktuellen Users vom Server ab
 // Erweitern um folgendes:
